+++ packages/zosfiles/__tests__/api/methods/download/Download.unit.test.ts
@@ -403,7 +403,7 @@ describe("z/OS Files - Download", () => {
             });
 
             downloadDatasetSpy.mockClear();
-            downloadDatasetSpy.mockImplementation(() => null);
+            downloadDatasetSpy.mockResolvedValue(null);
         });
 
         it("should throw and error if the data set name is not specified", async () => {
@@ -637,7 +637,7 @@ describe("z/OS Files - Download", () => {
             let caughtError;
 
             const dummyError = new Error("test");
-            listAllMembersSpy.mockImplementation(() => {
+            listAllMembersSpy.mockImplementation(async () => {
                 throw dummyError;
             });
 
@@ -661,7 +661,7 @@ describe("z/OS Files - Download", () => {
             let caughtError;
 
             const dummyError = new Error("test");
-            downloadDatasetSpy.mockImplementation(() => {
+            downloadDatasetSpy.mockImplementation(async () => {
                 throw dummyError;
             });
 
@@ -684,6 +684,40 @@ describe("z/OS Files - Download", () => {
             });
         });
 
+        it("should delay handling an error from Download.dataSet when failFast option is false", async () => {
+            let response;
+            let caughtError;
+
+            const dummyError = new Error("test");
+            downloadDatasetSpy.mockImplementation(async () => {
+                throw dummyError;
+            });
+
+            try {
+                response = await Download.allMembers(dummySession, dsname, {failFast: false});
+            } catch (e) {
+                caughtError = e;
+            }
+
+            const firstItem = listApiResponse.items[0];
+            const secondItem = listApiResponse.items[1];
+
+            expect(response).toBeUndefined();
+            expect(caughtError).toBeDefined();
+            expect(caughtError.message).toEqual(ZosFilesMessages.memberDownloadFailed.message +
+                `${firstItem.member.toLowerCase()}\n${secondItem.member.toLowerCase()}\n\n${dummyError.message}\n${dummyError.message}`);
+
+            expect(listAllMembersSpy).toHaveBeenCalledTimes(1);
+            expect(listAllMembersSpy).toHaveBeenCalledWith(dummySession, dsname, {});
+
+            expect(downloadDatasetSpy).toHaveBeenCalledTimes(2);
+            expect(downloadDatasetSpy).toHaveBeenCalledWith(dummySession, `${dsname}(${firstItem.member})`, {
+                file: `${dsFolder}/${firstItem.member.toLowerCase()}.txt`
+            });
+            expect(downloadDatasetSpy).toHaveBeenCalledWith(dummySession, `${dsname}(${secondItem.member})`, {
+                file: `${dsFolder}/${secondItem.member.toLowerCase()}.txt`
+            });
+        });
     });
 
     describe("datasetMatchingPattern", () => {